+++ source/utilities/detect-touch.js
@@ -1,3 +1,7 @@
+const IS_TOUCH = 'is-touch';
+const WITH_MOUSE = `${IS_TOUCH}--with-mouse`;
+const html = document.documentElement;
+
 class DetectTouch {
 
     get isTouchDevice() {
@@ -5,12 +9,28 @@ class DetectTouch {
     }
 
     constructor() {
-        this.touch = 'ontouchstart' in document.documentElement
+        this.hasMouse = false;
+        this.mouseEvent = () => this.handleMouseEvent();
+        this.touch = 'ontouchstart' in html
             || navigator.maxTouchPoints > 0
             || navigator.msMaxTouchPoints > 0;
 
+
+        html.addEventListener('touchstart', () => {
+            html.removeEventListener('mousemove', this.mouseEvent);
+        });
+
+        html.addEventListener('mousemove', this.mouseEvent);
+
         if (this.touch) {
-            document.querySelector('html').classList.add( 'is-touch' );
+            html.classList.add(IS_TOUCH);
+        }
+    }
+
+    handleMouseEvent() {
+        if(!this.hasMouse && html.classList.contains(IS_TOUCH)){
+            html.classList.add(WITH_MOUSE);
+            this.hasMouse = true;
         }
     }
 }